restructured linalg module
